+++ app/src/androidTest/java/com/graphhopper/navigation/example/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package graphhopper.com.graphhopper_navigation_example;
+package com.graphhopper.navigation.example;
 
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
