+++ app/src/main/java/com/graphhopper/navigation/example/NavigationLauncherActivity.java
@@ -1,4 +1,4 @@
-package graphhopper.com.graphhopper_navigation_example;
+package com.graphhopper.navigation.example;
 
 import android.app.DialogFragment;
 import android.content.Intent;
