+++ app/src/test/java/com/graphhopper/navigation/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package graphhopper.com.graphhopper_navigation_example;
+package com.graphhopper.navigation;
 
 import org.junit.Test;
