+++ build.gradle
@@ -19,9 +19,6 @@ allprojects {
     repositories {
         google()
         jcenter()
-        maven {
-            url "https://oss.sonatype.org/content/groups/public/"
-        }
     }
 }
